refactor: merge server into gateway — eliminate app/server package
Move deps.py, store.py, and LangGraph routers (threads, runs, assistants)
into app/gateway/. Delete app/server/ entirely.

All code now lives under one package: app/gateway/.
